backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Network Scanning
scapy==2.5.0

# Machine Learning
scikit-learn==1.3.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3

# Caching (optional)
redis==5.0.1

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
requests==2.31.0
